Add Textfile creation, misc changes
Feature Changes:
- Textfile now won't save if they have forward or backward slashes in their filenames
- Filepath string sanitation handles any amount of extra slashes, so you can slap 40 backslashes (or forward) in the location when filling the Textfile new/edit forms, and they will be ignored
- Errors caused by trying to save with invalid file name or path (e.g. using illegal characters as dictated by the OS) now handled

Housekeeping Changes:
- Moved predict_modifications method into new Helper class for organization purposes
- Removed redundant filepath name sanitation in Writer.rb
- Added a conditional check in Updater.rb to avoid unneeded looping when checking for external file renames/movements

+++ app/models/poros/writer.rb
@@ -1,24 +1,35 @@
 # Write (to) files based on edit forms
-# May involve renaming/deleting files
+# Involves renaming/deleting/creating files
+
+require 'fileutils'
 
 class Writer
 
-  def textfile(new_abs_path, old_abs_path)
-      new_name = new_abs_path.split("\\")[-1]
-      old_name = old_abs_path.split("\\")[-1]
-      construct = StringConstructor.new
-      new_pd = construct.parent_directory new_abs_path
-      old_pd = construct.parent_directory old_abs_path
+  def textfile(new_abs_path, old_abs_path=nil)
+    construct = StringConstructor.new
+    file = Scanner.new
+    # Duplicate used for argument because the original is frozen/immutable
+    new_name = new_abs_path.split("\\")[-1]
+    new_pd = construct.parent_directory new_abs_path
 
+    unless old_abs_path.nil?
+      old_name = old_abs_path.split("\\")[-1]
+      old_pd = construct.parent_directory old_abs_path
       if new_pd != old_pd
         # Instead of moving, delete and recreate
         File.delete(old_abs_path)
       elsif new_name != old_name
-        File.rename(pd + "\\" + old_name, new_abs_path)
+        begin
+          File.rename(old_abs_path, new_abs_path)
+        rescue Errno::EINVAL
+          return false
+        end
       end
+    end
 
-      tf = Textfile.find_by(location: new_abs_path)
-      Dir.mkdir new_pd unless File.exists? new_pd
+    tf = Textfile.find_by(location: new_abs_path)
+    begin
+      FileUtils::mkdir_p new_pd unless File.exists? new_pd
       File.open(new_abs_path, "w") do |f|
         # tf.contents has the correct number of CRLF characters,
         # => but the output file of f.write(tf.contents) has twice as many.
@@ -28,8 +39,13 @@ class Writer
         # => right number of newlines, while Notepad++ does not.
         # i.e I have no idea wtf is wrong, but this hack fixes it
         # => (for Notepad++ on Windows).
-        f.write(tf.contents.gsub("\r\n", "\n"))
+        contents = tf.contents.gsub("\r\n", "\n")
+        f.write(contents)
       end
+      true
+    rescue Errno::EINVAL
+      false
+    end
 
   end
